Changed to an Alternative version
In the original version there was no visible way of knowing if a thread has new posts etc.
This change doesn't fit a vanilla mybb install (there is a change to the forumbit_depth2_forum template in regards to the TD element width size)

Rather than hiding the original forum icons, this instead places the new custom icons into the TD background-image and then allows the original forum icons to appear over top (since their SPAN element is located above it).  Changing the original forum icons to use transparency and positioned correctly allows them to be shown on top of the custom forum images.